+++ src/renderer/void.rs
@@ -56,17 +56,7 @@ impl Renderer for Void {
         x: usize,
         y: usize,
     ) -> Result<(), ErrorKind> {
-        let size = data.dimensions();
-
-        if x + size.width > image.info.width() {
-            return Err(ErrorKind::ImageUpdateOutOfBounds);
-        }
-
-        if y + size.height > image.info.height() {
-            return Err(ErrorKind::ImageUpdateOutOfBounds);
-        }
-
-        Ok(())
+        data.check_update(&image.info, x, y)
     }
 
     fn delete_image(&mut self, image: Self::Image, _image_id: ImageId) {}